Change PlatformSettingsEditorUtility
