+++ Packages/UGF.EditorTools/Editor/IMGUI.PlatformSettings/PlatformSettingsEditorUtility.Deprecated.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 1d8e582d9538402aa584cb5a54af5db6
+timeCreated: 1601658029
